+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+[0.6.0]: https://github.com/AivGitHub/qworpa/releases/tag/v0.6.0
+
+## [0.6.0] - 2023-07-20
+
+### Added
+
+- Settings page.
+- User information in settings page.
+- Ability to change a password.
+
 [0.5.1]: https://github.com/AivGitHub/qworpa/releases/tag/v0.5.1
 
 ## [0.5.1] - 2023-07-19